% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Acoustic.R
\name{getAcousticData}
\alias{getAcousticData}
\title{Retrieve Acoustic data from the LifeWatch project}
\usage{
getAcousticData(startdate, stopdate, minband, maxband, by, params = FALSE, ...)
}
\arguments{
\item{startdate}{Starting date for the query in UTC. Must be a character date of the form "2020-02-26T10:35:00.00Z"}

\item{stopdate}{Stopping date for the query in UTC. Must be a character of the form "2020-02-26T11:35:00.00Z"}

\item{minband}{lower frequency band from which the sound value will be retrieved}

\item{maxband}{upper frequency band from which the sound value will be retrieved}

\item{by}{Sample period, one of "1 min", "1 hour" or "1 day"}

\item{params}{If TRUE, returns a list with the dataset and the query parameters applied in the server side. IF FALSE returns only the data.}

\item{...}{Params to be passed to lw_check_lwdataserver().}
}
\value{
Dataframe with the acoustic data within the specified daterange and location.
}
\description{
Retrieve Acoustic data from the LifeWatch project
}
\examples{
getAcousticData("2021-03-26T10:35:00.00Z", "2021-03-26T10:45:00.00Z",
                minband = 9, maxband = 11, by = "1 min")
}
