% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lwdataFunctions.R
\name{getBatsData}
\alias{getBatsData}
\title{Retrieve batcorder-data from the LifeWatch project}
\usage{
getBatsData(startdate, stopdate, by, params = FALSE)
}
\arguments{
\item{startdate}{Starting date for the query}

\item{stopdate}{Stopping date for the query}

\item{by}{Sample period, one of "1 min", "60 min" or "1 day"}

\item{params}{If TRUE, returns a list with the dataset and the query parameters applied in the server side. IF FALSE returns only the data.}
}
\value{
Dataframe with the aggregated batcorder-data within the specified daterange.
}
\description{
Retrieves the aggregated batcorder data from the LifeWatch project.
}
\examples{
getBatsData("2014-08-01", "2014-09-01", "1 min")
getBatsData("2014-08-01", "2014-09-01", "60 min", TRUE)
getBatsData("2014-08-01", "2014-09-01", "1 day", TRUE)
}
