% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lwdataFunctions.R
\name{getBuoyData}
\alias{getBuoyData}
\title{Retrieve Buoy-data from the LifeWatch project}
\usage{
getBuoyData(startdate, stopdate, stations, params = FALSE, ...)
}
\arguments{
\item{startdate}{Starting date for the query}

\item{stopdate}{Stopping date for the query}

\item{stations}{list of stations to be included in the query, currently list c("Buoy at C-Power","Spuikom Sluice","Buoy in Spuikom", "Ostend Research Tower"). Use \code{stations = "All"} to get all stations .}

\item{params}{If TRUE, returns a list with the dataset and the query parameters applied in the server side. IF FALSE returns only the data.}

\item{...}{Reserved for internal use.}
}
\value{
Dataframe with the buoy-data within the specified daterange and location.
}
\description{
Retrieves the aggregated buoy data from the LifeWatch project.
}
\examples{
getBuoyData("2021-03-19", "2021-04-21", "All")
getBuoyData("2021-03-19", "2021-04-21", "Buoy at C-Power", TRUE)
getBuoyData("2021-03-19", "2021-04-21", c("Spuikom Sluice", "Buoy in Spuikom",
"Ostend Research Tower"), TRUE)
}
