% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lwdataFunctions.R
\name{getFlowcamData}
\alias{getFlowcamData}
\title{Retrieve flowcam-data from the LifeWatch project}
\usage{
getFlowcamData(startdate, stopdate, params = FALSE)
}
\arguments{
\item{startdate}{Starting date for the query}

\item{stopdate}{Stopping date for the query}

\item{params}{If TRUE, returns a list with the dataset and the query parameters applied in the server side. IF FALSE returns only the data.}
}
\value{
Dataframe with the aggregated flowcam-data within the specified daterange.
}
\description{
Retrieves the aggregated flowcam data from the LifeWatch project.
}
\examples{
getFlowcamData("2020-04-19", "2020-04-21") # Only data
getFlowcamData("2020-04-19", "2020-04-21", TRUE) # Data + query parameters
}
