% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lwdataFunctions.R
\name{getMvbData}
\alias{getMvbData}
\title{Retrieve MVB-data from the LifeWatch project}
\usage{
getMvbData(
  startdate,
  stopdate,
  parameters,
  stations = NULL,
  by,
  calc,
  usr = NULL,
  pwd = NULL,
  params = FALSE
)
}
\arguments{
\item{startdate}{Starting date for the query. Without a login, this is restricted to the last month}

\item{stopdate}{Stopping date for the query. Without a login, this is restricted to the last month}

\item{parameters}{One (or list) of the \code{$parameter} in \code{\link{listMvbStations}}.}

\item{stations}{List of stations to be included in the query. See \code{\link{listMvbStations}}. Use \code{stations="All"} or \code{stations=NULL} to get all stations.}

\item{by}{Time grouping for calculation, one of ('day','hour', '10min' ,'none')}

\item{calc}{Calculation to perform given time grouping, one of ('avg', 'max', 'min', 'none').}

\item{usr}{Username to connect to database}

\item{pwd}{Password to connect to database}

\item{params}{If TRUE, returns a list with the dataset and the query parameters applied in the server side. IF FALSE returns only the data.}
}
\value{
Dataframe with the specified MVB data.
}
\description{
Retrieves the Meetnet Vlaams Banken (MVB) data from the LifeWatch project. Without valid credentials you are only allowed to view the "Tide TAW" parameter for the last 30 days, grouped by day/hour.
To get an account, register via the \href{http://rshiny.lifewatch.be/account?p=register}{Lifewatch RShiny registration} webpage.
}
\examples{
getMvbData(Sys.Date() - 30, Sys.Date() + 1, parameters = 'Tide TAW',
stations = "Nieuwpoort", by = "day", calc = "avg")
getMvbData(Sys.Date() - 30, Sys.Date() + 1, parameters = 'Tide TAW',
stations = "Blankenberge", by = "hour", calc = "max", params = TRUE)
}
