% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lwdataFunctions.R
\name{getStationData}
\alias{getStationData}
\title{Retrieve Station-data from the LifeWatch project}
\usage{
getStationData(
  startdate,
  stopdate,
  stations = "all",
  categories = "all",
  params = FALSE,
  ...
)
}
\arguments{
\item{startdate}{Starting date for the query}

\item{stopdate}{Stopping date for the query}

\item{stations}{list of stations to be included in the query. Use \code{stations="All"} to get all stations.}

\item{categories}{List of categories to return in query, one of  ("SPM", "CTD", "Nutrients", "Secchi", "Pigments"). Use \code{categories="All"} to get all categories}

\item{params}{If TRUE, returns a list with the dataset and the query parameters applied in the server side. IF FALSE returns only the data.}

\item{...}{Params to be passed to lw_check_lwdataserver().}
}
\value{
Dataframe with the Station-data.
}
\description{
Retrieves the station data from the LifeWatch project.
}
\examples{
getStationData("2019-07-22", "2019-07-23", stations = "all", categories = "all")
getStationData("2019-07-22", "2019-07-23", stations = c(120, 215),
categories = c("Nutrients", "Secchi"), params = TRUE)
}
