% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lwdataFunctions.R
\name{getUnderwayData}
\alias{getUnderwayData}
\title{Retrieve Underway-data from the LifeWatch project}
\usage{
getUnderwayData(startdate, stopdate, by, params = FALSE, ...)
}
\arguments{
\item{startdate}{Starting date for the query}

\item{stopdate}{Stopping date for the query}

\item{by}{Time grouping for data aggregation, one of ("1 day","60 min","10 min","1 min")}

\item{params}{If TRUE, returns a list with the dataset and the query parameters applied in the server side. IF FALSE returns only the data.}

\item{...}{Params to be passed to lw_check_lwdataserver().}
}
\value{
Dataframe with the specified data.
}
\description{
Retrieves the data collected by the MIDAS underway continuous sampling system on the RV Zeeleeuw and RV Simon Stevin as part of the LifeWatch project.
}
\examples{
getUnderwayData("2021-03-15", "2021-04-13", "1 day")
getUnderwayData("2021-03-15", "2021-04-13", "60 min", params = TRUE)
}
