% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lwdataFunctions.R
\name{getUvaBirdData}
\alias{getUvaBirdData}
\title{Retrieve UVA-bird-data from the LifeWatch project}
\usage{
getUvaBirdData(startdate, stopdate, tagcodes, by, usr = NULL,
  pwd = NULL, params = FALSE, ...)
}
\arguments{
\item{startdate}{Starting date for the query}

\item{stopdate}{Stopping date for the query}

\item{tagcodes}{List of tag numbers from \code{$tagnr} in \code{\link{listUvaTags}}. Can also use "All HG" "All LBB" or "All MH" (MH only when logged in).}

\item{by}{Time grouping for aggregation, one of ("1 day","60 min","10 min","1 min","1 week")}

\item{usr}{Username to connect to database}

\item{pwd}{Password to connect to database}

\item{params}{If TRUE, returns a list with the dataset and the query parameters applied in the server side. IF FALSE returns only the data.}

\item{...}{Params to be passed to lw_check_lwdataserver().}
}
\value{
Dataframe with the specified bird tracking data.
}
\description{
Retrieves the University of Amsterdam (UvA) Bird Tracking System data from the LifeWatch project.
Need valid authentication to access the entire data. To get an account, register via the \href{http://rshiny.lifewatch.be/account?p=register}{Lifewatch RShiny registration} webpage.
Without login you can only access data collected at least two years ago.
}
\examples{
getUvaBirdData("2013-06-10", "2013-06-14", tagcodes = c("719","6013","610"), by = "1 day")
getUvaBirdData("2013-06-10", "2013-06-14", tagcodes = "All HG", by = "1 week", params = TRUE)
}
