% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lwdataFunctions.R
\name{getZooscanData}
\alias{getZooscanData}
\title{Retrieve zooscan-data from the LifeWatch project}
\usage{
getZooscanData(startdate, stopdate, params = FALSE, ...)
}
\arguments{
\item{startdate}{Starting date for the query}

\item{stopdate}{Stopping date for the query}

\item{params}{If TRUE, returns a list with the dataset and the query parameters applied in the server side. IF FALSE returns only the data.}

\item{...}{Reserved for internal use.}
}
\value{
Dataframe with the aggregated zooscan-data within the specified daterange.
}
\description{
Retrieves the aggregated zooscan data from the LifeWatch project.
}
\examples{
getZooscanData("2011-01-01", "2021-04-14") # Only data
getZooscanData("2011-01-01", "2021-04-14", TRUE) # Data + query parameters
}
