% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lwdataFunctions.R
\name{listEtnProjects}
\alias{listEtnProjects}
\title{Retrieve an overview of the ETN data.}
\usage{
listEtnProjects(usr = NULL, pwd = NULL, ...)
}
\arguments{
\item{usr}{Username to connect to ETN database}

\item{pwd}{Password to connect to ETN database}

\item{...}{params to be passer to lw_check_lwdataserver}
}
\value{
Dataframe with name and type of networks.
}
\description{
Retrieves the name of network/animals?
To get an account, register via the \href{http://rshiny.lifewatch.be/account?p=register}{Lifewatch RShiny registration} webpage.
}
\examples{
listEtnProjects()
}
