% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lwdataFunctions.R
\name{listUvaTags}
\alias{listUvaTags}
\title{List available  UVA bird tags}
\usage{
listUvaTags(usr = NULL, pwd = NULL)
}
\arguments{
\item{usr}{Username to connect to ETN database}

\item{pwd}{Password to connect to ETN database}
}
\description{
To get an account, register via the \href{http://rshiny.lifewatch.be/account?p=register}{Lifewatch RShiny registration} webpage.
}
