% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lwdataFunctions.R
\name{lwdata}
\alias{lwdata}
\title{Function to retrieve data from the LifeWatch project}
\usage{
lwdata(
  datatype = "Buoy data",
  from = as.character(Sys.Date() - 90),
  to = as.character(Sys.Date()),
  stations = c("Buoy at C-Power"),
  binSize = "60 min",
  calc = "Time bins",
  UrlPar = NULL,
  code = NULL,
  posres = 2,
  logged = FALSE,
  projectlist = NULL,
  tagprojectlist = NULL,
  loggedInUserPostgresUsername = NULL,
  loggedInUserPostgresPwd = NULL,
  phylasp = "#2#51#",
  taxranks = c(Species = 220),
  qualities = c("Hi"),
  processing = "Validated",
  params = FALSE,
  force_opencpu = FALSE
)
}
\arguments{
\item{datatype}{Type of data to request}

\item{from}{Starting date for the query}

\item{to}{Stopping date for the query}

\item{stations}{(Buoy, MVB, Station) list of stations to be included in the query}

\item{binSize}{(Bats, ETN, CPOD, MVB, UvaBird, Underway) Sample period.}

\item{calc}{(MVB) Calculation to perform given time grouping, one of ('avg', 'max', 'min', 'none')}

\item{UrlPar}{.}

\item{code}{.}

\item{posres}{.}

\item{logged}{(CPOD, MVB, ETN, UvaBird) Request data under moratory if you have an account}

\item{projectlist}{.}

\item{tagprojectlist}{.}

\item{loggedInUserPostgresUsername}{(CPOD, MVB, ETN, UvaBird) Postgres username}

\item{loggedInUserPostgresPwd}{(CPOD, MVB, ETN, UvaBird) Postgres password}

\item{phylasp}{.}

\item{taxranks}{.}

\item{qualities}{.}

\item{processing}{(CPOD) One of ('Validated','Raw')}

\item{params}{If TRUE, returns a list with the dataset and the query parameters applied in the server side. IF FALSE returns only the data.}

\item{force_opencpu}{boolean. If true, the query will be forced to run through the OpenCPU server.}
}
\value{
The sum of \code{x} and \code{y}
}
\description{
Large wrapper function that retrieves the LifeWatch data. Use datatype specific functions for clearer use.
}
\examples{
lwdata()
lwdata('listETNprojects')
lwdata('zooscan-data')
lwdata('flowcam-data')
lwdata('MVB')
lwdata('station-data')
lwdata('cpod-data')
lwdata('etn-data')
}
