% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lwdataFunctions.R
\name{lwdata2}
\alias{lwdata2}
\title{For use in the LifeWatch Rshiny data-explorer (LWDE). Reads input and USER arguments}
\usage{
lwdata2(type, input = NULL, USER = NULL, ...)
}
\arguments{
\item{type}{Datatype as defined in LWDE}

\item{input}{LWDE input element, as a list. Can be obtained by using shiny::reactiveValuesToList()}

\item{USER}{LWDE USER element, as a list. Can be obtained by using shiny::reactiveValuesToList()}

\item{...}{Params to be passed to lw_check_lwdataserver()}
}
\value{
Returns dataframe of the requested datatype
}
\description{
For use in the LifeWatch Rshiny data-explorer (LWDE). Reads input and USER arguments
}
\examples{
lwdata2("listETNprojects")
input=c()
input$daterange = c(Sys.Date()-365*2, Sys.Date())
lwdata2("ZooScan data", input)
lwdata2("flow", input)
}
