% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lwdataFunctions.R
\name{getCpodData}
\alias{getCpodData}
\title{Retrieve C-POD data from the LifeWatch project}
\usage{
getCpodData(
  startdate,
  stopdate,
  processing = "Raw",
  quality = c("Hi", "Mod", "Lo"),
  by = "1 min",
  usr = NULL,
  pwd = NULL,
  params = FALSE,
  projectcode = NULL,
  species = NULL,
  ...
)
}
\arguments{
\item{startdate}{Starting date for the query}

\item{stopdate}{Stopping date for the query}

\item{processing}{Data processing, currently only "Raw".}

\item{quality}{One or more of ("Hi","Mod", "Lo").}

\item{by}{Sample period, currently only "1 min"}

\item{usr}{Username to connect to ETN database}

\item{pwd}{Password to connect to ETN database}

\item{params}{If TRUE, returns a list with the dataset and the query parameters applied in the server side. IF FALSE returns only the data.}

\item{projectcode}{One or more of ("Apelafico_acoustics", "CODEVCO", "Lifewatch_extra", "Lifewatch_test", "PAM-Borssele", "PelFish", "PhD_Parcerisas", "PureWind", "STRAITS_PAM", "cpod-lifewatch", "cpod-od-natuur")}

\item{species}{One or more of ("Dolphins","NBHF","sonar").}

\item{...}{Reserved for internal use.}
}
\value{
Dataframe with the specified C-POD data.
}
\description{
Retrieves the C-POD (Cetacean Acoustic Hydrophone Network) data from the LifeWatch project.
Need valid authentication to access the entire data.To get an account, register via the \href{http://rshiny.lifewatch.be/account?p=register}{Lifewatch RShiny registration} webpage.
}
\examples{
getCpodData("2020-04-20", "2020-04-21", quality = c("Hi", "Lo"), params = TRUE)
getCpodData("2020-04-20", "2020-04-21", quality = c("Hi", "Lo"), params = FALSE, species = c("Dolphins", "NBHF"), projectcode = "cpod-lifewatch")
}
