% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lwdataFunctions.R
\name{getCpodData}
\alias{getCpodData}
\title{Retrieve C-POD data from the LifeWatch project}
\usage{
getCpodData(
  startdate,
  stopdate,
  processing,
  quality,
  by,
  usr = NULL,
  pwd = NULL,
  params = FALSE
)
}
\arguments{
\item{startdate}{Starting date for the query}

\item{stopdate}{Stopping date for the query}

\item{processing}{One of ('Validated','Raw')}

\item{quality}{One of ("Hi","Mod", "Lo", "Hi+Mod") .....}

\item{by}{Sample period, one of ("1 day", "1 week", "60 min","10 min", "1 min")}

\item{usr}{Username to connect to ETN database}

\item{pwd}{Password to connect to ETN database}

\item{params}{If TRUE, returns a list with the dataset and the query parameters applied in the server side. IF FALSE returns only the data.}
}
\value{
Dataframe with the specified C-POD data.
}
\description{
Retrieves the C-POD (Cetacean Acoustic Hydrophone Network) data from the LifeWatch project.
Need valid authentication to access the entire data.To get an account, register via the \href{http://rshiny.lifewatch.be/account?p=register}{Lifewatch RShiny registration} webpage.
}
\examples{
getCpodData("2020-04-19", "2020-04-21", processing = "Validated", quality = "Hi+Mod", by = "1 week")
getCpodData("2020-04-19", "2020-04-21", processing = "Raw", 
quality = c("Hi", "Lo"), by = "1 day", params = TRUE)
}
