% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTD.R
\name{getCTDData}
\alias{getCTDData}
\title{Retrieve CTD data from the LifeWatch project}
\usage{
getCTDData(startdate, stopdate, stations = "All", by, params = FALSE, ...)
}
\arguments{
\item{startdate}{Starting date for the query in UTC. Must be a character date of the form "YYYY-MM-DD".}

\item{stopdate}{Stopping date for the query in UTC. Must be a character date of the form "YYYY-MM-DD".}

\item{stations}{List of stations to be included in the query. Use stations = "All" to get all stations. See stations list in details section.}

\item{by}{Aggregation criteria. One of c("Per meter", "Per cast")}

\item{params}{If TRUE, returns a list with the dataset and the query parameters applied in the server side. IF FALSE returns only the data.}

\item{...}{Reserved for internal use.}
}
\value{
Dataframe with the CTD data within the specified daterange and location.
}
\description{
Retrieve CTD data from the LifeWatch project
}
\details{
List of stations:
c("120", "130", "230", "215", "330", "421", "435", "710", "700",
"780", "ZG02", "W09", "W08", "LW01", "W10", "LW02", "W07bis")
}
\examples{
getCTDData("2023-01-15", "2023-02-15", stations = c("120", "W09"), by = "Per meter")
getCTDData("2023-01-15", "2023-02-15", stations = "All", by = "Per cast", params = TRUE)
}
